Modified: changed users_signin to users_api, moved new user registration, tests in users_api.
