--- mbtb_app/resources/apis/users/register/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class RegisterConfig(AppConfig):
-    name = 'register'
