--- mbtb_app/resources/apis/users/register/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path, include
-from . import views
-from rest_framework import routers
-
-router = routers.DefaultRouter()
-router.register('add_new_users', views.NewUsersViewSet)
-
-urlpatterns = [
-    path('', include(router.urls)),
-]
