+++ mbtb_app/resources/apis/users/users_api/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class UsersSigninConfig(AppConfig):
-    name = 'users_signin'
+    name = 'users_api'
