+++ mbtb_app/resources/apis/users/users_api/urls.py
@@ -3,6 +3,7 @@ from . import views
 from rest_framework import routers
 
 router = routers.DefaultRouter()
+router.register('add_new_users', views.NewUsersViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
