--- mbtb_app/resources/apis/users/users_signin/models.py
@@ -1,3 +1,0 @@
-from django.db import models
-
-# Create your model here.
